Renamed designer into robot_designer
